hosts/zabbix-server/opt/zabbix-discord/requirements.txt
disnake==2.4.0
py-zabbix==1.1.7
python-dotenv==0.19.2
aiozabbix==1.1.2
python-dateutil==2.8.2

